Task1

+++ Task1/src/main/java/org/task1/Developer.java
@@ -1,7 +1,7 @@
 package org.task1;
 
 public class Developer extends Employee {
-    public Developer(String name, String surname, Float baseSalary, Integer experience) {
+    public Developer(String name, String surname, Double baseSalary, Integer experience) {
         super(name, surname, baseSalary, experience);
     }
 }
