+++ Task1/src/main/java/org/task1/Employee.java
@@ -3,7 +3,7 @@ package org.task1;
 import java.math.BigDecimal;
 
 public class Employee {
-    public Employee(String name, String surname, Float baseSalary, Integer experience) {
+    public Employee(String name, String surname, Double baseSalary, Integer experience) {
         this.name = name;
         this.surname = surname;
         this.baseSalary = baseSalary;
@@ -12,14 +12,14 @@ public class Employee {
 
     public String name;
     public String surname;
-    public Float baseSalary;
+    public Double baseSalary;
     public Integer experience;
-    public Float countedSalary() {
+    public Double countedSalary() {
         if (experience >= 2 && experience < 5) {
             return baseSalary + 200;
         }
         if (experience >= 5) {
-            return baseSalary * (float)1.2 + 500;
+            return baseSalary * (Double)1.2 + 500;
         }
         return baseSalary;
     }
